Fixes/refactoring as per suggestions
