+++ tests/cbmc/proofs/s2n_socket_is_ipv6/s2n_socket_is_ipv6_harness.c
@@ -19,5 +19,5 @@ void s2n_socket_is_ipv6_harness()
 
   /* Post-condition. */
   assert(S2N_IMPLIES(result == S2N_SUCCESS, (ipv6 != NULL && (*ipv6 == 0 || *ipv6 == 1))));
-  assert(S2N_IMPLIES(ipv6 == NULL, result < S2N_SUCCESS));
+  assert(S2N_IMPLIES(ipv6 == NULL, result != S2N_SUCCESS));
 }
