+++ tests/cbmc/proofs/s2n_socket_read/Makefile
@@ -14,11 +14,11 @@ REMOVE_FUNCTION_BODY +=
 UNWINDSET +=
 
 PROOF_SOURCES += $(PROOFDIR)/$(HARNESS_FILE)
+PROOF_SOURCES += $(PROOF_SOURCE)/make_common_datastructures.c
+PROOF_SOURCES += $(PROOF_STUB)/getpeername.c
 PROOF_SOURCES += $(PROOF_STUB)/read.c
 PROOF_SOURCES += $(PROOF_STUB)/s2n_calculate_stacktrace.c
-PROOF_SOURCES += $(PROOF_STUB)/cbmc_allocate_s2n_socket_read_io_context.c
 
-PROJECT_SOURCES += $(SRCDIR)/utils/s2n_result.c
 PROJECT_SOURCES += $(SRCDIR)/utils/s2n_socket.c
 
 include ../Makefile.common
