+++ tests/cbmc/proofs/s2n_socket_read_restore/s2n_socket_read_restore_harness.c
@@ -7,10 +7,10 @@
  */
 
 #include <utils/s2n_socket.h>
+#include <cbmc_proof/make_common_datastructures.h>
 
 void s2n_socket_read_restore_harness()
 {
-  struct s2n_socket_read_io_context *cbmc_allocate_s2n_socket_read_io_context();
   /* Non-deterministic inputs. */
   struct s2n_connection *s2n_connection = malloc(sizeof(*s2n_connection));
   if (s2n_connection != NULL) {
