+++ tests/cbmc/proofs/s2n_socket_was_corked/Makefile
@@ -14,8 +14,8 @@ REMOVE_FUNCTION_BODY +=
 UNWINDSET +=
 
 PROOF_SOURCES += $(PROOFDIR)/$(HARNESS_FILE)
+PROOF_SOURCES += $(PROOF_SOURCE)/make_common_datastructures.c
 PROOF_SOURCES += $(PROOF_STUB)/s2n_calculate_stacktrace.c
-PROOF_SOURCES += $(PROOF_STUB)/cbmc_allocate_s2n_socket_write_io_context.c
 
 PROJECT_SOURCES += $(SRCDIR)/utils/s2n_socket.c
